query_with_nested_images/requirements.txt
langchain-community
langchain
langchain-openai
pymilvus
numpy 
streamlit
pandas
torchvision
torch
openpyxl
python-dotenv
opencv-python
langchain-huggingface
langchain-chroma
PyMuPDF
segment-anything
pillow
easyocr
streamlit-option-menu
watchdog
langchain_milvus
langchain_experimental
tabulate
psycopg2